% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCellsByCluster.R
\name{dataMatrixToCluster.pca}
\alias{dataMatrixToCluster.pca}
\title{dataMatrixToCluster.pca is a function to cluster a data matrix
by t-SNE and return an annotated list of cells}
\usage{
dataMatrixToCluster.pca(df, cols.to.exclude = c())
}
\description{
dataMatrixToCluster.pca is a function to cluster a data matrix
by t-SNE and return an annotated list of cells
}
