% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCellsByCluster.R
\name{dataMatrixToCluster.seurat}
\alias{dataMatrixToCluster.seurat}
\title{dataMatrixToCluster.seurat is a function to cluster a data matrix
by seurat and return an annotated list of cells}
\usage{
dataMatrixToCluster.seurat(df, cols.to.exclude = c())
}
\description{
dataMatrixToCluster.seurat is a function to cluster a data matrix
by seurat and return an annotated list of cells
}
