% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCellsByCluster.R
\name{dataMatrixToCluster.tsne}
\alias{dataMatrixToCluster.tsne}
\title{dataMatrixToCluster.tsne is a function to cluster a data matrix
by t-SNE and return an annotated list of cells}
\usage{
dataMatrixToCluster.tsne(df, cols.to.exclude = c())
}
\description{
dataMatrixToCluster.tsne is a function to cluster a data matrix
by t-SNE and return an annotated list of cells
}
