% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneCellLists.R
\name{getGeneList}
\alias{getGeneList}
\title{getGeneList grabs the gene list from teh synapse table}
\usage{
getGeneList(method = "cibersort")
}
\description{
getGeneList grabs the gene list from teh synapse table
}
