% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAnalysis.R
\name{launchApp}
\alias{launchApp}
\title{launchApp is a function to prouducea shiny app for a dataste}
\usage{
launchApp(sample.mat, cell.annotations)
}
\description{
launchApp is a function to prouducea shiny app for a dataste
}
