% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneCellLists.R
\name{plotGeneListByCluster}
\alias{plotGeneListByCluster}
\title{get list of clusters, annotate by}
\usage{
plotGeneListByCluster(gene.annotations, samples, cell.annotations)
}
\description{
get list of clusters, annotate by
}
