% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAnalysis.R
\name{runMarkdown}
\alias{runMarkdown}
\title{runMarkdown is a function to prouduce rMd analysis files}
\usage{
runMarkdown(samp.mat, cell.annotations, syn_file, analysis_dir)
}
\description{
runMarkdown is a function to prouduce rMd analysis files
}
